Foundation of ILL-Impl
Using elements of RecCalc to implement the syntax, pretty printer, and parser for ILL-Impl. Parser just has the lexer and reservedOps as of this commit. Pretty does not implement the logical operations promote, discard, copy, derelict yet. Syntax has a small issue if whether we need a tensor term AND a tensor type.